feat(auth): add login schema (#10)
### TL;DR

This PR introduces a new login form and schema for validating inputs.

### What changed?

Two new files were added: login-form.tsx that implements the form and login-schema.ts which defines the format that will be used for validation of inputs.

### How to test?

- Pull the changes
- Navigate to the login form in the front-end application
- Try to submit with invalid inputs and check if the validations take place

### Why make this change?

The login form and schema are essential components of the authentication process in the app. It helps to ensure users' data are correctly entered and validated before submission.

---