TIM-105 feat(backend): add newsfeed & connections table (#41)
### TL;DR

Added two new tables 'newsfeed' and 'connections' to the database.

### What changed?

The created 'newsfeed' table stores the content posted by users, associated with user id, and time of creation. The 'connections' table keeps track of the established connections between users. User access is controlled through appropriate security policies.

### How to test?

Check if the tables have been created in the database and the security policies are in place. Validate if new entries are being added correctly in the 'newsfeed' and 'connections' tables.

### Why make this change?

These changes were necessary to facilitate the creation of posts by users and track their connections.

---