TIM-108 feat(auth): add loading state to login form (#53)
### TL;DR

This Pull Request improves the user experience of the login form.

### What changed?

Two changes were made to the login form. The input fields (email and password) are now disabled while the login request is being processed to prevent any changes. Secondly, the login button now shows a spinner and is disabled during login requests to show the user that their request is being processed.

### How to test?

Check out the PR branch and navigate to the login form. Attempt to login and observe that the input fields and login button get disabled and a spinner shows on the login button until the request is processed.

### Why make this change?

This change improves the user feedback during the login process by showing a spinner on the login button and disabling the input fields. This prevents any confusion for the users regarding the state of their login request.

---