TIM-155 feat(newsfeed): add function to fetch newsfeed-likes (#109)
### TL;DR

This PR introduces new utility functions for formatting and counting the likers of a post, as well as fetching the likers from the newsfeed.

### What changed?

Two new files have been added: `format-likers.ts` in the home/feed directory which includes functions for formatting liker data, and `get-newsfeed_likes.ts` in the queries directory for fetching likes associated with a newsfeed post. Added functionality to handle 'unknown' liker, list two likers by first name, and count extra likers if there are more than two.

### How to test?

Test these by checking the formatted data for likers on the newsfeed likes, check the responses from the `getNewsfeedLikes` query.

### Why make this change?

This change provides necessary functionalities for a robust and dynamic display of 'likes' on a post in the newsfeed

---